chore: Move HomePage.tsx from /pages to /features/home/
- Moved the HomePage.tsx file to /features/home/ to align with a modern React feature-based folder structure.
