+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import './App.css';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import HomePage from './pages/Home';
+import HomePage from './features/home/Home';
 import RouteLayout from './pages/RootLayout';
 import ErrorPage from './pages/ErrorPage';
 import BookDetails from './pages/BookDetails';
